+++ config/config.ex
@@ -2,6 +2,7 @@ import Config
 
 config :request_cache,
   enabled?: true,
+  verbose?: false,
   graphql_paths: ["/graphiql", "/graphql"],
   conn_priv_key: :__shared_request_cache__,
   request_cache_module: RequestCache.ConCacheStore,